Allows multi-p to be assigned to need
Summary:
- Separated out tests based on components of the coordinator
- Allow multiple people to be assigned (and able to participate) in a
need -- dont block if someone is already on the need!
- TODO: Allow user to be assigned to many Half Half Needs (rather than
excluding users who are already assigned to other parts of the
incident?)
