+++ imports/api/OpportunisticCoordinator/server/coordinator.tests.js
@@ -5,8 +5,7 @@ import {
   updateAvailability
 } from './identifier';
 import {CONSTANTS} from "../../Testing/testingconstants";
-import {createIncidentFromExperience, startRunningIncident,
-        setIntersection, sustainedAvailabilities} from "../../OCEManager/OCEs/methods";
+import {createIncidentFromExperience, startRunningIncident} from "../../OCEManager/OCEs/methods";
 import {Experiences, Incidents} from "../../OCEManager/OCEs/experiences";
 import { Submissions } from "../../OCEManager/currentNeeds";
 import {insertTestOCE } from "../populateDatabase";
@@ -208,350 +207,4 @@ describe('Assignments Collection test', () => {
     }
   });
 
-});
-
-describe('test checkIfThreshold. Single Need, Single UID; allowRepeatContributions: false', () => {
-
-  const incident_id = Random.id();
-  const eid = Random.id();  // doesn't really matter
-  const userA = Random.id();
-  const NEEDNAME = 'Coffee Time';
-  const updatedIncidentsAndNeeds = [
-    {
-      _id: incident_id,
-      needUserMaps: [
-        {
-          needName: NEEDNAME,
-          users: [
-            {uid: userA, place: 'placeA', distance: 10.0}
-          ]
-        }
-      ]
-    }
-  ];
-  const numberNeeded = 4;
-  beforeEach(() => {
-    resetDatabase();
-
-    Incidents.insert({
-      _id: incident_id,
-      eid: eid,
-      callbacks: null, // dont need callbacks
-      contributionTypes: [
-        {
-          needName: NEEDNAME,
-          situation: {
-            detector: Random.id(),
-            number: 1
-          },
-          numberNeeded: numberNeeded,
-          notificationDelay: 1,
-          // not including the parameter defaults to false
-          // allowRepeatContributions: false
-        }
-      ]
-    });
-
-    // userA is available for incident.need1
-    Availability.insert({
-      _id: incident_id,
-      needUserMaps: [
-        {
-          needName: NEEDNAME, users: [
-            { uid: userA, place: "place1", distance: 10.0 }
-          ]
-        },
-      ],
-    });
-
-    // userA NOT assigned to incident yet
-    // runNeedsWithThresholdMet does these types of updates via adminUpdatesForAddingUserToIncident
-    // and the call to this function comes after checkIfThreshold
-    Assignments.insert({
-      _id: incident_id,
-      needUserMaps: [
-        { needName: NEEDNAME, users: [] },
-      ],
-    });
-
-    // Empty submissions ready to be filled
-    for (let i = 0; i < numberNeeded; ++i) {
-      Submissions.insert({
-        _id: Random.id(),
-        eid : Random.id(),
-        iid : incident_id,
-        needName : NEEDNAME,
-        uid : null,
-      });
-    }
-  });
-
-  it('user SHOULD be able to participate on the first try', () => {
-    let incidentsWithUsersToRun = checkIfThreshold(updatedIncidentsAndNeeds);
-
-    // should look something like this
-    // { vPnAsWkhjv8EN6n9p: { 'Coffee Time': [ 'tDm59tFq2XBBKQZm5' ] } }
-    chai.assert.isNotNull(incidentsWithUsersToRun, 'incidentsWithUsersToRun should not be empty');
-    chai.assert.isNotNull(incidentsWithUsersToRun[incident_id], 'incidentsWithUsersToRun should contain incident');
-    chai.assert.isNotNull(incidentsWithUsersToRun[incident_id][NEEDNAME], 'incidentsWithUsersToRun should contain needName');
-    let users_for_need = incidentsWithUsersToRun[incident_id][NEEDNAME];
-    let foundUser = users_for_need.find((userMeta) => userMeta.uid === userA);
-    chai.assert(foundUser, 'incidentsWithUsersToRun should contain userA');
-  });
-
-  it('user SHOULD NOT be allowed to participate twice', () => {
-    // But user has already participated in the past
-    Submissions.insert({
-      _id: Random.id(),
-      eid : Random.id(),
-      iid : incident_id,
-      needName : NEEDNAME,
-      uid : userA,
-      content : {
-        "this": "is a previous submission"
-      }
-    });
-
-    let incidentsWithUsersToRun = checkIfThreshold(updatedIncidentsAndNeeds);
-
-    // should look something like this
-    // { vPnAsWkhjv8EN6n9p: {} }
-    chai.assert.isNotNull(incidentsWithUsersToRun, 'incidentsWithUsersToRun should not be empty');
-    chai.assert.isNotNull(incidentsWithUsersToRun[incident_id], 'incidentsWithUsersToRun should contain incident');
-
-    // object should be empty
-    let obj = incidentsWithUsersToRun[incident_id];
-    if (Object.keys(obj).length === 0 && obj.constructor === Object) {
-      // TODO(rlouie): do this with multiple needs going on, or multiple users
-      chai.assert(true, 'incidentsWithUsersToRun should NOT contain needName or the User');
-    } else {
-      chai.assert(false, 'incidentsWithUsersToRun should NOT contain needName or the User');
-    }
-  });
-
-});
-
-describe('test checkIfThreshold; Single Need, Single UID; allowRepeatContributions: true', () => {
-
-  const incident_id = Random.id();
-  const eid = Random.id();  // doesn't really matter
-  const userA = Random.id();
-  const NEEDNAME = 'Coffee Time';
-  const updatedIncidentsAndNeeds = [
-    {
-      _id: incident_id,
-      needUserMaps: [
-        {
-          needName: NEEDNAME,
-          users: [
-            {uid: userA, place: 'placeA', distance: 10.0}
-          ]
-        }
-      ]
-    }
-  ];
-  const numberNeeded = 4;
-  beforeEach(() => {
-    resetDatabase();
-
-    Incidents.insert({
-      _id: incident_id,
-      eid: eid,
-      callbacks: null, // dont need callbacks
-      contributionTypes: [
-        {
-          needName: NEEDNAME,
-          situation: {
-            detector: Random.id(),
-            number: 1
-          },
-          numberNeeded: numberNeeded,
-          notificationDelay: 1,
-          allowRepeatContributions: true
-        }
-      ]
-    });
-
-    // userA is available for incident.need1
-    Availability.insert({
-      _id: incident_id,
-      needUserMaps: [
-        {
-          needName: NEEDNAME,
-          users: [
-            {uid: userA, place: 'placeA', distance: 10.0}
-          ]
-        },
-      ],
-    });
-
-    // userA NOT assigned to incident yet
-    // runNeedsWithThresholdMet does these types of updates via adminUpdatesForAddingUserToIncident
-    // and the call to this function comes after checkIfThreshold
-    Assignments.insert({
-      _id: incident_id,
-      needUserMaps: [
-        { needName: NEEDNAME, users: [] },
-      ],
-    });
-
-    // Empty submissions ready to be filled
-    for (let i = 0; i < numberNeeded; ++i) {
-      Submissions.insert({
-        _id: Random.id(),
-        eid : Random.id(),
-        iid : incident_id,
-        needName : NEEDNAME,
-        uid : null,
-      });
-    }
-  });
-
-  it('user SHOULD be able to participate on the first try', () => {
-    let incidentsWithUsersToRun = checkIfThreshold(updatedIncidentsAndNeeds);
-
-    // should look something like this
-    // { vPnAsWkhjv8EN6n9p: { 'Coffee Time': [ 'tDm59tFq2XBBKQZm5' ] } }
-    chai.assert.isNotNull(incidentsWithUsersToRun, 'incidentsWithUsersToRun should not be empty');
-    chai.assert.isNotNull(incidentsWithUsersToRun[incident_id], 'incidentsWithUsersToRun should contain incident');
-    chai.assert.isNotNull(incidentsWithUsersToRun[incident_id][NEEDNAME], 'incidentsWithUsersToRun should contain needName');
-    let users_for_need = incidentsWithUsersToRun[incident_id][NEEDNAME];
-    let foundUser = users_for_need.find((userMeta) => userMeta.uid === userA);
-    chai.assert(foundUser, 'incidentsWithUsersToRun should contain userA');
-  });
-
-  it('user ALSO SHOULD be allowed to participate twice', () => {
-    // But user has already participated in the past
-    Submissions.insert({
-      _id: Random.id(),
-      eid : Random.id(),
-      iid : incident_id,
-      needName : NEEDNAME,
-      uid : userA,
-      content : {
-        "this": "is a previous submission"
-      }
-    });
-
-    let incidentsWithUsersToRun = checkIfThreshold(updatedIncidentsAndNeeds);
-
-    // should look something like this
-    // { vPnAsWkhjv8EN6n9p: { 'Coffee Time': [ 'tDm59tFq2XBBKQZm5' ] } }
-    chai.assert.isNotNull(incidentsWithUsersToRun, 'incidentsWithUsersToRun should not be empty');
-    chai.assert.isNotNull(incidentsWithUsersToRun[incident_id], 'incidentsWithUsersToRun should contain incident');
-    chai.assert.isNotNull(incidentsWithUsersToRun[incident_id][NEEDNAME], 'incidentsWithUsersToRun should contain needName');
-    let users_for_need = incidentsWithUsersToRun[incident_id][NEEDNAME];
-    let foundUser = users_for_need.find((userMeta) => userMeta.uid === userA);
-    chai.assert(foundUser, 'incidentsWithUsersToRun should contain userA');
-  });
-
-});
-
-describe('Sustained (place, need) Match for Availability Dictionary', function() {
-  let availabilityDictionary = {
-    "asianFoodCrawlIncident": [
-      ['ramen_dojo', 'noodleNeed', 10.0],
-      ['kongs_chinese', 'noodleNeed', 20.5]
-    ],
-    "groceryBuddiesIncident": [
-      ['trader_joes', 'groceryNeed', 20.0]
-    ],
-    "sunsetTogether": [
-      ['', 'sunsetTogetherNeed', undefined],
-      ['ramen_dojo', 'sunsetTogetherNeed', undefined],
-      ['kongs_chinese', 'sunsetTogetherNeed', undefined],
-      ['trader_joes', 'sunsetTogetherNeed', undefined]
-    ]
-  };
-
-  // sustained success - after notification delay
-  let sustainedAfterAvailDict = {
-    "asianFoodCrawlIncident": [
-      ['ramen_dojo', 'noodleNeed', 3.0]
-    ]
-  };
-
-  // sustained for not place need -- after notification delay
-  let sustainedWeatherTimeNeedAfterAvailDict = {
-    "sunsetTogether": [
-      ['', 'sunsetTogetherNeed', undefined],
-    ]
-  };
-
-  // not sustained - after notification delay
-  let notSustainedAfterAvailDict = {
-    "asianFoodCrawlIncident": [
-      ['onsen_oden', 'noodleNeed', 25.0]
-    ]
-  };
-
-  it('Sustained Incident', function() {
-    let incidentIntersection = setIntersection(Object.keys(availabilityDictionary), Object.keys(sustainedAfterAvailDict));
-    console.log(incidentIntersection);
-    chai.assert.equal(
-      JSON.stringify(incidentIntersection),
-      JSON.stringify(["asianFoodCrawlIncident"]));
-  });
-
-  it('Sustained [Place, Needs]', function() {
-    let incident = "asianFoodCrawlIncident";
-
-    let beforePlacesAndNeeds = availabilityDictionary[incident].map((place_need_dist) => place_need_dist.slice(0,2));
-    let afterPlacesAndNeeds = sustainedAfterAvailDict[incident].map((place_need_dist) => place_need_dist.slice(0,2));
-    let place_need_intersection = setIntersection(
-      beforePlacesAndNeeds, afterPlacesAndNeeds);
-
-    chai.assert.equal(
-      JSON.stringify(place_need_intersection),
-      JSON.stringify([['ramen_dojo', 'noodleNeed']])
-    )
-  });
-
-  // different places - after notification delay
-  it('NOT Sustained [Place, Needs]', function() {
-    let incident = "asianFoodCrawlIncident";
-
-    let place_need_intersection = setIntersection(
-      availabilityDictionary[incident],
-      notSustainedAfterAvailDict[incident]);
-
-    chai.assert.equal(
-      JSON.stringify(place_need_intersection),
-      JSON.stringify([])
-    )
-  });
-
-  it('Sustained AvailDict', function() {
-    let sustainedAvailDict = sustainedAvailabilities(availabilityDictionary, sustainedAfterAvailDict);
-    chai.assert.equal(
-      JSON.stringify(sustainedAvailDict),
-      JSON.stringify({
-        "asianFoodCrawlIncident": [
-          ['ramen_dojo', 'noodleNeed', 3.0]
-        ]
-      })
-    )
-  });
-
-  it('NOT Sustained AvailDict', function() {
-    let sustainedAvailDict = sustainedAvailabilities(availabilityDictionary, notSustainedAfterAvailDict);
-    chai.assert.equal(
-      JSON.stringify(sustainedAvailDict),
-      JSON.stringify({})
-    );
-
-    chai.assert.equal(Object.keys(sustainedAvailDict).length, 0);
-  });
-
-  it('Sustained NonPlace AvailDict', function() {
-    let sustainedAvailDict = sustainedAvailabilities(availabilityDictionary, sustainedWeatherTimeNeedAfterAvailDict);
-    chai.assert.equal(
-      JSON.stringify(sustainedAvailDict),
-      JSON.stringify({
-        "sunsetTogether": [
-          ['', 'sunsetTogetherNeed', undefined]
-        ]
-      })
-    );
-  });
 });
